refactor(projects): simplify project creation page and enhance loading states
- Replaced the existing project creation form with a dynamic CreateProjectForm component for improved modularity.
- Integrated loading spinner for better user experience during data fetching.
- Updated project schema to handle geospatial data and file references.
- Enhanced form validation to include new location and file handling requirements.
