+++ components/features/projects/index.ts
@@ -5,4 +5,5 @@ export { default as ProjectCard } from './ProjectCard'
 export { default as ProjectStats } from './ProjectStats'
 export { default as ProjectFilters } from './ProjectFilters'
 export { default as ProjectsPage } from './ProjectsPage'
+export { default as CreateProjectForm } from './CreateProjectForm'
 export type { ProjectFilters as ProjectFiltersType } from './ProjectFilters'
