+++ lib/database/schemas/projects.ts
@@ -16,7 +16,7 @@ export const projectSchema = z.object({
   budget: validationPatterns.positiveNumber,
   category: z.array(tradeCategoryEnum).default([]),
   pid: validationPatterns.nonEmptyString,
-  location: validationPatterns.nonEmptyString,
+  location: validationPatterns.geospatialLocation, // Updated to handle geospatial data
   certificate_of_title: validationPatterns.optionalUrl,
   project_type: projectTypeEnum,
   status: projectStatusEnum.default('Draft'),
@@ -26,8 +26,8 @@ export const projectSchema = z.object({
   expiry_date: z.date(),
   substantial_completion: validationPatterns.optionalDate,
   is_verified_project: z.boolean().default(false),
-  project_photos: z.array(validationPatterns.url).default([]),
-  files: z.array(validationPatterns.url).default([]),
+  project_photos: z.array(validationPatterns.fileReference).default([]), // Updated to handle actual files
+  files: z.array(validationPatterns.fileReference).default([]), // Updated to handle actual files
   creator: validationPatterns.uuid,
   proposal_count: z.number().int().min(0).default(0),
 })
@@ -37,5 +37,7 @@ export type ProjectType = z.infer<typeof projectTypeEnum>
 export type ProjectStatus = z.infer<typeof projectStatusEnum>
 export type VisibilitySettings = z.infer<typeof visibilitySettingsEnum>
 export type TradeCategory = z.infer<typeof tradeCategoryEnum>
+export type GeospatialLocation = z.infer<typeof validationPatterns.geospatialLocation>
+export type FileReference = z.infer<typeof validationPatterns.fileReference>
 
 export const validateProject = (data: unknown): Project => projectSchema.parse(data)
